Make templates HTML5-conformant
If no doctype is specified, browsers fall back to quirks mode. However,
"charset" attribute used in <meta> by templates was unknown in that time
of ancients. So a proper doctype is specified to eliminate contradiction
leading to nasty browser-specific behavior.

Also, a neat title was added to each template to be HTML5-conformant.